refactor: some naming refactoring and destructurize
