+++ src/App.tsx
@@ -2,7 +2,7 @@ import { MantineProvider } from '@mantine/core';
 import Routing from './components/Routing/Routing';
 import { NotesProvider } from './context/NotesContext';
 import { ModalsProvider } from '@mantine/modals';
-import { AuthProvider } from './context/AuthProvider';
+import { AuthProvider } from './context/AuthContext';
 import '@mantine/tiptap/styles.css';
 import '@mantine/core/styles.css';
 import '@mantine/spotlight/styles.css';
