+++ src/components/AuthStatus/AuthStatus.tsx
@@ -1,14 +1,14 @@
 import { useNavigate } from 'react-router-dom';
-import { useAuthContext } from '@/context/AuthProvider';
+import { useAuthContext } from '@/context/AuthContext';
 import { Box, Button, Flex, Text } from '@mantine/core';
 import { AppRoutes } from '@/types/generalTypes';
 
 export const AuthStatus = () => {
-  const auth = useAuthContext();
+  const { signOut, user } = useAuthContext();
   const navigate = useNavigate();
 
   const handleSignOut = () => {
-    auth?.signOut(() => navigate(AppRoutes.Main));
+    signOut(() => navigate(AppRoutes.Main));
   };
 
   function EnterBtn() {
@@ -31,7 +31,7 @@ export const AuthStatus = () => {
     <Flex>
       {/* <Flex rowGap='sm' direction={'column'} justify='center' align='flex-end'> */}
       {/* if user is logged in */}
-      {auth?.user && (
+      {user && (
         <>
           <Box w={125}>
             <Text
@@ -42,14 +42,14 @@ export const AuthStatus = () => {
               truncate='end'
               variant='gradient'
               gradient={{ from: 'grape', to: 'red', deg: 90 }}>
-              {auth?.user}
+              {user}
             </Text>
           </Box>
           <ExitBtn />
         </>
       )}
       {/* if user is not logged in */}
-      {!auth?.user && <EnterBtn />}
+      {!user && <EnterBtn />}
     </Flex>
   );
 };
