--- src/components/TextEditor/index.ts
@@ -1,3 +1,0 @@
-import { TextEditorToolbar } from './TextEditorToolbar';
-
-export { TextEditorToolbar };
